fix: structure/cleanup
